+++ test/lua/test_veafInterpreter.lua
@@ -18,6 +18,7 @@ local _base = debug.getinfo(1, "S").source:match("^@(.+)[\\/]") or "."
 luaunit = dofile(_base .. "/luaunit.lua") -- exported as global for test methods
 dofile(_base .. "/dcs_mocks.lua")
 dofile(_base .. "/../../src/scripts/veaf/veaf.lua")
+dofile(_base .. "/../../src/scripts/veaf/veafScheduler.lua")
 dofile(_base .. "/../../src/scripts/veaf/veafInterpreter.lua")
 
 -- ---------------------------------------------------------------------------